refactor: 범위 밖 숫자 검증 기능을 Lotto 내부로 이동
- 테스트 기능 리팩토링 추가
