+++ src/main/java/lotto/model/Lotto.java
@@ -1,10 +1,13 @@
 package lotto.model;
 
+import static lotto.Constraints.MAX_NUMBER;
+import static lotto.Constraints.MIN_NUMBER;
 import static lotto.Constraints.SIZE;
 
 import java.util.List;
 import lotto.model.exceptions.DuplicatedNumberException;
 import lotto.model.exceptions.InvalidSizeException;
+import lotto.model.exceptions.OutOfRangeNumberException;
 
 public class Lotto {
     private final List<Integer> numbers;
@@ -17,6 +20,7 @@ public class Lotto {
     private void validate(List<Integer> numbers) {
         validateSize(numbers);
         validateNumbersDuplicate(numbers);
+        validateInRangeNumber(numbers);
     }
 
     // 6자리 검증
@@ -31,6 +35,15 @@ public class Lotto {
         if (numbers.stream().distinct().count() != SIZE.getValue()) {
             throw new DuplicatedNumberException();
         }
+    }
+
+    // 범위 밖 숫자 검증
+    private void validateInRangeNumber(List<Integer> numbers) {
+        numbers.stream()
+                .filter(number -> number > MAX_NUMBER.getValue() || number < MIN_NUMBER.getValue())
+                .forEach(number -> {
+                    throw new OutOfRangeNumberException();
+                });
     }
 
     public boolean contains(int number) {
